requirements.txt
streamlit==1.32.0
google-genai==0.3.0
pandas==2.2.1
sodapy==2.2.0
python-dotenv==1.0.1